Time: 33 ms (84.33%) | Memory: 61.7 MB (69.40%) - LeetSync

+++ GreedyAlgorithms/2681-put-marbles-in-bags/put-marbles-in-bags.java
@@ -7,9 +7,9 @@ class Solution {
      */
     public long putMarbles(int[] weights, int k) {
         int n = weights.length;
-        long[] pairSum = new long[n - 1]; // SC: O(N)
+        int[] pairSum = new int[n - 1]; // SC: O(N)
         for (int i = 1; i < n; i++) { // TC: O(N)
-            pairSum[i - 1] = (long) weights[i - 1] + (long) weights[i];
+            pairSum[i - 1] = weights[i - 1] + weights[i];
         }
         Arrays.sort(pairSum); // TC: O(N x Log(N))
         // we need to pick the (k - 1) pairs for k bags
